+++ ramp.cpp
@@ -118,7 +118,7 @@ void Ramp::WriteRegDefaults()
    SetRegValueInt(strKeyName,"RampType", m_d.m_type);
    SetRegValue(strKeyName,"TimerEnabled",REG_DWORD,&m_d.m_tdr.m_fTimerEnabled,4);
    SetRegValue(strKeyName,"TimerInterval",REG_DWORD,&m_d.m_tdr.m_TimerInterval,4);
-   SetRegValue(strKeyName,"Image", REG_SZ, &m_d.m_szImage, strlen(m_d.m_szImage));
+   SetRegValue(strKeyName,"Image", REG_SZ, &m_d.m_szImage, (DWORD)strlen(m_d.m_szImage));
    SetRegValue(strKeyName,"ImageMode",REG_DWORD,&m_d.m_imagealignment,4);
    SetRegValueBool(strKeyName,"ImageWalls",m_d.m_fImageWalls);
    SetRegValueBool(strKeyName,"CastsShadow",m_d.m_fCastsShadow);
